+++ CHANGELOG.md
@@ -8,4 +8,10 @@ complete example.
 fix the c++ header include problem.
 
 ## 0.0.7
-statusEventChannel and dataEventChannel 
+statusEventChannel and dataEventChannel
+
+## 0.0.8
+fix compile error
+
+## 0.0.9
+netCallback all on main thread